@@ -1,7 +1,7 @@
 source 'http://rubygems.org'
 ruby '2.3.3'
 
-gem 'rails', '= 5.0.0'
+gem 'rails', '< 5.1'
 
 gem 'citrus', '2.4.1'
 gem 'coffee-rails'
