@@ -8,7 +8,12 @@ class NestedReferenceDecorator < ReferenceDecorator
 
     # Fall back to nesting reference's PDF is nestee does not have one.
     def pdf_link
-      super unless reference.downloadable?
+      return nesting_reference_pdf_link unless reference.downloadable?
       helpers.link_to 'PDF', reference.url
+    end
+
+    def nesting_reference_pdf_link
+      return unless nesting_reference.downloadable?
+      helpers.link_to 'PDF', nesting_reference.url
     end
 end
